requirements.txt
flask
gradio
tensorflow
numpy
pandas
scikit-learn
requests
pyserial
